# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Rokwire Groups Building Block API Documentation.",
        "title": "Rokwire Groups Building Block API",
        "contact": {},
        "version": "1.29.0"
    },
    "host": "localhost",
    "basePath": "/gr",
    "paths": {
...
        "/api/admin/group/{groupID}/posts": {
            "get": {
                "security": [
                    {
                        "AppUserAuth": []
                    }
                ],
                "description": "gets all posts for the desired group.",
                "tags": [
                    "Admin"
                ],
                "operationId": "AdminGetGroupPosts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "APP",
                        "name": "APP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "groupID",
                        "name": "groupID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Values: message|post",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "asc|desc",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Post"
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/group/{groupId}/posts": {
            "post": {
                "security": [
                    {
                        "AppUserAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ],
                "description": "creates a post within the desired group.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "operationId": "AdminCreateGroupPost",
                "parameters": [
                    {
                        "type": "string",
                        "description": "APP",
                        "name": "APP",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Post"
                        }
                    }
                }
            }
        },
        "/api/admin/group/{groupId}/posts/{postId}": {
            "get": {
                "security": [
                    {
                        "AppUserAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ],
                "description": "Gets a post within the desired group.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "operationId": "AdminGetGroupPost",
                "parameters": [
                    {
                        "type": "string",
                        "description": "APP",
                        "name": "APP",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Post"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "AppUserAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ],
                "description": "Updates a post within the desired group.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "operationId": "AdminUpdateGroupPost",
                "parameters": [
                    {
                        "type": "string",
                        "description": "APP",
                        "name": "APP",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Post"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "AppUserAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ],
                "description": "Updates a post within the desired group.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "operationId": "AdminDeleteGroupPost",
                "parameters": [
                    {
                        "type": "string",
                        "description": "APP",
                        "name": "APP",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/analytics/posts": {
            "get": {
                "security": [
                    {
                        "IntAPIKeyAuth": []
                    }
                ],
                "description": "Gets posts",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "operationId": "AnalyticsGetPosts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "group_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start date string - RFC3339 encoded",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date string - RFC3339 encoded",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/rest.analyticsGetPostsResponse"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/group/{groupID}/posts": {
            "get": {
                "security": [
                    {
                        "AppUserAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ],
                "description": "gets all posts for the desired group.",
                "tags": [
                    "Client"
                ],
                "operationId": "GetGroupPosts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "APP",
                        "name": "APP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "groupID",
                        "name": "groupID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Values: message|post",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "asc|desc",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Post"
                            }
                        }
                    }
                }
            }
        },
        "/api/group/{groupId}/posts": {
            "post": {
                "security": [
                    {
                        "AppUserAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ],
                "description": "creates a post within the desired group.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "operationId": "CreateGroupPost",
                "parameters": [
                    {
                        "type": "string",
                        "description": "APP",
                        "name": "APP",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Post"
                        }
                    }
                }
            }
        },
        "/api/group/{groupId}/posts/{postId}": {
            "get": {
                "security": [
                    {
                        "AppUserAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ],
                "description": "Gets a post within the desired group.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "operationId": "GetGroupPost",
                "parameters": [
                    {
                        "type": "string",
                        "description": "APP",
                        "name": "APP",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Post"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "AppUserAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ],
                "description": "Updates a post within the desired group.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "operationId": "UpdateGroupPost",
                "parameters": [
                    {
                        "type": "string",
                        "description": "APP",
                        "name": "APP",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Post"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "AppUserAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ],
                "description": "Updates a post within the desired group.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "operationId": "DeleteGroupPost",
                "parameters": [
                    {
                        "type": "string",
                        "description": "APP",
                        "name": "APP",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/groups/{id}": {
            "get": {
                "security": [
                    {
                        "AppUserAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ],
                "description": "Gives a group",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "operationId": "GetGroup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "APP",
                        "name": "APP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/getGroupResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "AppUserAuth": []
                    }
                ],
                "description": "Updates a group.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Client"
                ],
                "operationId": "UpdateGroup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "APP",
                        "name": "APP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/updateGroupRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Creator": {
            "type": "object",
            "required": [
                "email",
                "name",
                "user_id"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "ToMember": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "external_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "model.Post": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "org_id": {
                    "type": "string"
                },
                "date_created": {
                    "type": "string"
                },
                "date_notified": {
                    "type": "string"
                },
                "date_scheduled": {
                    "type": "string"
                },
                "date_updated": {
                    "type": "string"
                },
                "group_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "is_abuse": {
                    "type": "boolean"
                },
                "member": {
                    "$ref": "#/definitions/Creator"
                },
                "parent_id": {
                    "type": "string"
                },
                "private": {
                    "type": "boolean"
                },
                "reactions": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "replies": {
                    "description": "This is constructed by the code (ParentID)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Post"
                    }
                },
                "subject": {
                    "type": "string"
                },
                "to_members": {
                    "description": "nil or empty means everyone; non-empty means visible to those user ids and admins",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ToMember"
                    }
                },
                "top_parent_id": {
                    "type": "string"
                },
                "use_as_notification": {
                    "type": "boolean"
                }
            }
        },
...
        "rest.analyticsGetPostsResponse": {
            "type": "object",
            "properties": {
                "org_id": {
                    "type": "string"
                },
                "date_created": {
                    "type": "string"
                },
                "date_updated": {
                    "type": "string"
                },
                "group_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "member_user_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "APIKeyAuth": {
            "type": "apiKey",
            "name": "ROKWIRE-API-KEY",
            "in": "header"
        },
        "AppUserAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header (add Bearer prefix to the Authorization value)"
        },
        "IntAPIKeyAuth": {
            "type": "apiKey",
            "name": "INTERNAL-API-KEY",
            "in": "header"
        }
    }
}